Add a helper function to provide reasonable defaults
The `KeySize` should be at least 2048bits, which is the hard lower bound for keysize and must be raised if 2048bit-RSA-keys considered to weak. However one may provide a custom lower bound as default value.
So first check if the default value is lower than the global minimum and raise, if it’s too low.
Return the given value or the default value, whichever is higher.

Signed-off-by: Mathias Kaufmann <me@stei.gr>