pkg/controller: log correct renewal time
This fixes a case where a negative time would be logged.
Fixes #341
It fixes that issue by never renewing a certificate more quickly than
30s, even if it's already expired.

In addition, it also logs a more accurate time by reworking the
scheduled work queue to return the actual time the item will be
processed in.

This differs from before since previously, the rate limiting time would
not be known ahead of time and so couldn't be included in any logged
information.

To plumb that information up, it was necessary to include the rate
limiter directly in the ScheduledWorkQueue rather than using the
'AddRateLimited' layer of abstraction it was before.