Update troubleshooting.rst
The example for checking the namespace has the correct label was incorrect.

It was also being hidden on the Github RST rendering, because of the `:emphasize-lines`. I suspect this is why the mistake has never been noticed previously.
I suggest having the example render in github is more important the emphasizing the line.