Fix address to reference array element
The address of a for loop created variable does not change between loops - the variable is over-written with each new variable.  This fixes the code so that `d` references the array element instead of the local variable, which means that when `d` is assigned to `matchAll` or `specificMatch` it will work as intended.

Signed-off-by: Dobes Vandermeer <dobesv@gmail.com>