Avoid using cert name in Secret / Order labels
In the previous version, each Secret and Order would get a label with
the full name of the owning Certificate.  Because labels are limited to
63 chars, certificate names beyond that size would cause a number of
problems (e.g. certificates not getting renewed.)

For Secrets, the Certificate name is already being stored as an
annotation, so this patch removes the label altogether.  The annotation
may be used to retrieve the Certificate name from the secret when
needed.  This change is applied at
pkg/cainjector/controller:OwningCertForSecret as part of the patch.

For Orders, the Certificate is linked via an OwnerReference, which can
be used to filter.

Signed-off-by: Galo Navarro <anglorvaroa@gmail.com>