Add flags to configure cainjector leader election
This commit enables users to customize the behavior of the cainjector
regarding leader election by adding `--leader-election-lease-duration`,
`--leader-election-renew-deadline` and `--leader-election-retry-period`
flags. These flags are already available on the controller.

Prior to this change, the values of these options were inherited from
controller-runtime. This commit keeps the same defaults for the
cainjector, for backward compatibility, even though they differ from
those of the controller.

Signed-off-by: Julien ALEXANDRE <22328659+ndrpnt@users.noreply.github.com>