Add leader election flags in Helm chart
This commit lets Helm chart users set the value of the
`--leader-election-lease-duration`, `--leader-election-renew-deadline`
and `--leader-election-retry-period` flags. These flags are configured
globally for all components, to be consistent with the rest of the
chart.

Since their default values differ in the controller and the cainjector,
they are unset by default in the Helm chart, for backward compatibility.

Signed-off-by: Julien ALEXANDRE <22328659+ndrpnt@users.noreply.github.com>