e2e: rename "validation" to "sanity-check"
There is a confusion around the word "validation" since we often use it
in the context of "verify the X.509 certificate". To avoid confusion and
since these tests never actually verify the X.509 certificate (due to
the fact that we do not have access to the root CA generated by Pebble),
I propose to rename "validations" with "sanity check".

Signed-off-by: Maël Valais <mael@vls.dev>