Issue #3853
cert-manager create the HTTP01 ACME solver pod, when can specify:

The image: --acme-http01-solver-image
The serviceAccount (##3817): --acme-http01-solver-service-account
The resources
But we can't specify the imagePullSecret to use.

This is an issue for:

Air gapped environments with private registries
Docker Hub rate limits for anonymous pulls
It is currently possible to use the serviceAccount with an attached
imagePullSecret, but it is not always convenient because it requires to
update the serviceAccount of every namespace in the cluster to link the
imagePullsecret.

Signed-off-by: primael <primael.bruant@gmail.com>