remove unused ValidateIssuedCertificate function
this function doesn't appear to be used anywhere, which makes it a
little confusing when trying to work out how the checks are done in the
e2e tests.

given that we encourage people not to import cert-manager as a module
and that anyone who does is likely not to use functions in the test
directory, it seems safe enough to remove this rather than deprecating
it.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>