remove krew automation
github actions is pretty flawed in several ways, and it's missing
features which would make it usable for us.

we want to trigger our krew automation when a new stable release of our
latest supported cert-manager release is published. github has all this
information, but it's not easy to _use_ that information when triggering
a github action.

we also only want to trigger the action on the main repo and not on
forks.

if we end up having to write custom code to do this, we might as well
add that code into cmrel and not have to deal with the flaws in gh
actions.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>