issuing: reproduce bug where issuing controller picks up oudated req
In 6484010f I made sure that the issuing controller would not use the
status of an outdated CertificateRequest for setting its own status. By
"outdated", I mean that the CertificateRequest was created at a
moment when the Certificate has a different spec.

The bug fix I introduced in 6484010f is only able to detect changes in
the Certificate resource that can be compared to the CertificateRequest.
That includes the following fields in the Certificate:

- spec.commonName
- spec.dnsNames
- spec.ipAddresses
- spec.uris
- spec.emailAddresses
- spec.subject
- spec.isCA
- spec.keyUsages
- spec.duration
- spec.issuerRef

But some fields on the Certificate can't be compared to anything on the
CertificateRequest, making it impossible for the issuing controller to
know when it should not bubble up the CR status into the Certificate.
The information that can't be detected are:

- spec.privateKey.rotationPolicy
- spec.privateKey.encoding
- spec.privateKey.algorithm
- spec.privateKey.size
- spec.encodeUsagesInRequest

Note: some Certificate fields do not need to be compared since they do
not need a re-issuance of the certificate when they are changed:

- spec.keystores.jks
- spec.keystores.pkcs12
- spec.renewBefore
- spec.revisionHistoryLimit

Signed-off-by: Maël Valais <mael@vls.dev>