issuing controller: move a one-time-called test func to where it is called
In order to work around an import cycle when importing MustCreateCryptoBundle
from the tests in the "certificates" package, I propose to move the function
GenerateLocallySignedTemporaryCertificate to the only place it is used: the
issuing_controller_test.go file.

I also made it private since this function seems to be very specific to the
issuing controller and (I think) won't be useful anywhere else.

Signed-off-by: Maël Valais <mael@vls.dev>