make: rebuild images when a different commit is checked out
When switching branches, the Go files may not change. But since the
images contain the commit hash, e.g.:

  cert-manager-controller-amd64:v1.7.0-beta.0-142-gfc0819af6

It is surprising when trying to deploy to Kind: the git commit that is
checked out does not match the commit hash of the image.

To avoid confusion, I added bin/release-version that gets updated only
when the currently checked out commit changes.

Signed-off-by: Maël Valais <mael@vls.dev>