[release-1.8] Use variable for curl, add retries
This adds multiple retries on every attempt we make to use curl, which
should help to reduce flakes. Uses a $(CURL) variable where possible so
that we have the same invocation everywhere.

Also switches to using the more verbose curl arguments, in an attempt to
make it easier to reason about how curl is configured.

The backport of this was manual because of conflicts relating to
$(BINDIR), and our new methods of installing kubebuilder tools and the
gatewayapi code

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>