Allow ECDSA for ACME client keys
Lego uses ec256 by default, which I hadn't realized previously, so
when I asked Let's Encrypt for a bump on limits on an account, it was
using ECDSA. The problem is I can't use this account in cert-manager,
since it's explicitly rejected.

golang/x/crypto already accepts EDCSA anyway, so the limitation is
solely in cert-manager.

https://pkg.go.dev/golang.org/x/crypto/acme#Client
https://github.com/go-acme/lego/blob/0e614c1d0bd56884d6e4c8a05f97e4e653bf8db8/cmd/flags.go#L49
Signed-off-by: James M Leddy <james.leddy@mongodb.com>