add trivy scan targets
These enable scanning of each of our container images on linux/amd64
to check for vulnerabilities. These targets can then be used in CI as
an indicator that we might need to take a look at upgrading dependencies
or base images.

This was manually edited compared with the commit on master which this
was cherry-picked from; that involved "undoing" the BINDIR changes
present in the cherry-picked commit; we still use "bin" in release-1.8

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>