bump version of contour helm chart to 10.0.1
This matches the version used on the master branch at the time of this
commit. This version was chosen because it was tested on master. It
implies a bump of the underlying version of contour from 1.20 -> 1.23,
but it seems better to use a tested version of the chart rather than a
previously untested version.

This update is required because the previous version was removed from
the bitnami helm chart repo.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>